feat: Remove upper limit for IntelliJ Platform version
The changes in this commit remove the `pluginUntilBuild` property from the
`gradle.properties` file. This property was used to limit the compatibility of
the plugin to a specific range of IntelliJ Platform versions. By removing this
property, the plugin will now be compatible with all future versions of the
IntelliJ Platform, making it more future-proof and easier to maintain.

Increases the plugin version from 2.4.7 to 2.4.8 in the gradle.properties file.
This is a minor version increase as it includes new features or improvements,
but no breaking changes.

fixes: https://github.com/nEdAy/Flutter-Toolkit/issues/24